fix: make a value's shape a property of its column, not of its record (#101)
Issue #101 scanned every normalized string and number in the corpus and
reported five classes of value-shaped defect. Re-measured against the
corpus as it stands, three of them are real and two are the scan's own
heuristic. Both findings are here, because the ones that are not defects
cost as much to rediscover as the ones that are.

What was wrong, and is not any more:

* **Money was `int` here and `float` there.** `_parse_eur_amount` and its
  two era siblings returned an `int` whenever the printed figure happened
  to be whole, so 225 money columns held both types and 4.2 million values
  sat in one. Money is a float everywhere now; `values.as_money` is the
  one place that says so.
* **17,894 money values were still strings** — `id001s[].sandorio-suma`
  (8,658), `id001s[].sandorio-suma-lt` (7,047) and
  `vii-sandoriai[].suma-skaiciais` (2,189), against a documented invariant
  that money is never a string. The currency is load-bearing on the first
  (8,111 euro and 547 litas in one column) and is now a sibling key,
  `sandorio-suma-valiuta`. The value-band codes that look numeric stay
  strings: `"001"` is a band, and its leading zeros are the band.
* **Two columns are headed with a pair of names.** VRK heads the 2007-2008
  gift and service columns `Dovana, data` and `Paslauga, data` and prints
  both values in the one cell; slugified they were indistinguishable from
  the corpus's 21 other date keys. 129 gift and 24 service descriptions
  were hiding in a field named for a date, and each is now the pair it
  names. Ten cells printing the same date twice hold it once.
* **3,856 values across 30 elections ended in a bare `,` or `;`.**
* **30 biographies were a single NUL byte** rendered as `U+FFFD`, and read
  as text rather than as the nothing they are.

What is not a defect, measured:

* The "20,704 glued run-ons" are VRK's. 835 of the junctions checked
  against the retained HTML: 797 verbatim, no tag boundary between the two
  letters. 13,158 of the 20,651 are `VšĮ`; the rest are company names and
  VRK's own typing (`kAUNO`, `Partija tTvarka`). Inserting spaces there
  would corrupt 13,000 institution names to fix nothing.
* The `U+FFFD` cannot be re-decoded. Fetching a 2004 page live returns the
  replacement character in VRK's own bytes. Where a `"` closes the phrase
  the pair is unambiguous and the opening `„` is restored (153 of 173);
  the other 20 are left as published.

One shared primitive, `scraper/shared/values.py`, holds the rules, and the
era normalizers call it rather than growing a copy — a rule wired into six
normalizers and not the seventh is exactly the failure this repository
keeps paying for.

Chasing the fix across the corpus turned up a gap in the gate itself:
`reparse_diff.py --full` fell back to the fixture tree only for an election
with no retained tree at all, so 96 candidates across ten elections with a
tracked fixture and no retained page could never be written by `--apply`
while the fixture run went on reporting them as drifted. The fallback is
per candidate now, and those ten elections re-parse at 100% coverage.

Corpus: 112,046 records re-parsed and applied, 96 more reached by the new
fallback, and the 38 whose page exists in neither tree brought forward by
`scripts/backfill_value_hygiene.py`. All three gates green — 0 records
differ, no coverage findings, nothing new against the anomaly baseline.
`tests/test_corpus_value_hygiene.py` walks every record and pins the
invariants, plus the two counts that cannot go to zero, so a rise in
either is a finding.

+++ tests/test_savivaldybiu_1997_anketa_parser.py
@@ -174,13 +174,15 @@ class DeclarationTests(unittest.TestCase):
         self.assertEqual(self.declaration["valiuta"], "Lt")
 
     def test_the_combined_section_totals_are_present(self):
+        # Float, not int: issue #101 made money a property of the column
+        # rather than of whether this candidate's figure happened to be whole.
         for key in (
             "turtas-ir-pinigines-lesos-metu-pradzioje",
             "turtas-ir-pinigines-lesos-metu-pabaigoje",
             "kalendoriniais-metais-isigytas-turtas",
         ):
             with self.subTest(key):
-                self.assertIsInstance(self.declaration[key], int)
+                self.assertIsInstance(self.declaration[key], float)
 
     def test_the_modern_split_keys_are_null_not_invented(self):
         self.assertIsNone(self.declaration["privalomas-registruoti-turtas"])
